PACS/ndicom_server/requirements.txt
Django==2.2
djangorestframework==3.8.2
jsonfield==2.0.2
numpy==1.17.3
opencv_python==3.4.8.29
dipy==1.1.1
pynetdicom==1.5.0.dev0
Django==2.2.13
django_jsonfield==1.4.0
djangorestframework==3.11.0
Pillow==7.1.2
scikit_learn==0.23.1
google-pasta==0.1.6
opencv-python==3.4.8.29
psycopg2-binary==2.8.5
pytz==2018.4
cython
scikit-learn==0.22.2
tornado==5.0.2

